requirements.txt
pytest>=7.0.0
pytest-html>=3.2.0
pytest-timeout>=2.1.0
pytest-rerunfailures>=11.0
pytest-xdist>=3.0.0
paramiko>=2.11.0
requests>=2.28.0
flask>=2.2.0
pyyaml>=6.0
jinja2>=3.1.0
grpcio>=1.50.0
grpcio-tools>=1.50.0
